requirements.txt
cherrypy==18.8.0
cloudevents==1.8.0
ConfigArgParse==1.5.3
isodate==0.6.1
lxml==4.9.1
opentelemetry-api==1.15.0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-sdk==1.15.0
pylast==5.1.0
pytz==2022.6
rabe-cridlib==0.3.0
requests==2.28.1
Werkzeug==2.2.2
